+++ src/discof/restore/utils/fd_sshttp.c
@@ -540,9 +540,9 @@ read_response( fd_sshttp_t * http,
   int               minor_version;
   int               status;
   const char *      message;
-  ulong             message_len;
+  size_t            message_len;
   struct phr_header headers[ 128UL ];
-  ulong             header_cnt = 128UL;
+  size_t            header_cnt = 128UL;
   int parsed = phr_parse_response( http->response,
                                     http->response_len,
                                     &minor_version,
